# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RPG World Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface WorldState {
  worldName: string;
  currentEra: string;
  currentLocation: string;
  weather: string;
  timeOfDay: string;
  /**
   * @minimum 0
   * @maximum 10
   */
  tension: number;
  characterCount: number;
  locationCount: number;
  factionCount: number;
  itemCount: number;
  turnCount: number;
  lastUpdated: string;
}

export type TurnWorldUpdates = { [key: string]: unknown };

export interface Turn {
  id: number;
  sessionId: string;
  turnNumber: number;
  playerAction: string;
  narrative: string;
  worldUpdates?: TurnWorldUpdates;
  createdAt: string;
}

export interface SessionLog {
  sessionId: string;
  turns: Turn[];
  startedAt: string;
  worldName: string;
}

export interface PlayerTurn {
  /** The player's action or dialogue */
  action: string;
  /** The session ID to continue */
  sessionId: string;
}

export type TurnResultWorldUpdates = { [key: string]: unknown };

export interface TurnResult {
  turn: Turn;
  worldUpdates?: TurnResultWorldUpdates;
  updatedEntities: string[];
}

export interface NewSessionRequest {
  worldName?: string;
  playerCharacterName?: string;
  playerCharacterClass?: string;
  /** Optional seed description for the world */
  worldSeed?: string;
}

export type CharacterRole = (typeof CharacterRole)[keyof typeof CharacterRole];

export const CharacterRole = {
  player: "player",
  npc: "npc",
  enemy: "enemy",
  ally: "ally",
} as const;

export type CharacterStatus =
  (typeof CharacterStatus)[keyof typeof CharacterStatus];

export const CharacterStatus = {
  alive: "alive",
  dead: "dead",
  unknown: "unknown",
  missing: "missing",
} as const;

export interface Character {
  id: number;
  name: string;
  role: CharacterRole;
  class?: string;
  race?: string;
  level?: number;
  health?: number;
  maxHealth?: number;
  currentLocation?: string;
  description?: string;
  traits?: string[];
  status: CharacterStatus;
  updatedAt: string;
}

export interface Location {
  id: number;
  name: string;
  type: string;
  description: string;
  region?: string;
  discovered: boolean;
  /**
   * @minimum 0
   * @maximum 10
   */
  danger?: number;
  notableFeatures?: string[];
  updatedAt: string;
}

export interface Faction {
  id: number;
  name: string;
  description: string;
  alignment?: string;
  /**
   * @minimum 0
   * @maximum 10
   */
  power?: number;
  /**
   * @minimum -10
   * @maximum 10
   */
  playerRelation?: number;
  goals?: string[];
  updatedAt: string;
}

export type ItemType = (typeof ItemType)[keyof typeof ItemType];

export const ItemType = {
  weapon: "weapon",
  armor: "armor",
  potion: "potion",
  artifact: "artifact",
  misc: "misc",
  key: "key",
} as const;

export type ItemRarity = (typeof ItemRarity)[keyof typeof ItemRarity];

export const ItemRarity = {
  common: "common",
  uncommon: "uncommon",
  rare: "rare",
  legendary: "legendary",
  artifact: "artifact",
} as const;

export interface Item {
  id: number;
  name: string;
  type: ItemType;
  description: string;
  rarity: ItemRarity;
  ownedBy?: string;
  location?: string;
  magical: boolean;
  updatedAt: string;
}