# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RPG World Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /session:
    get:
      operationId: getSessionLog
      tags: [session]
      summary: Get session log
      description: Returns the current session log with all turns
      responses:
        "200":
          description: Session log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionLog"
    post:
      operationId: processTurn
      tags: [session]
      summary: Process a player turn
      description: Submits a player action and returns the DM narrative with world updates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerTurn"
      responses:
        "200":
          description: Turn result with narrative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /session/new:
    post:
      operationId: startNewSession
      tags: [session]
      summary: Start a new session
      description: Starts a fresh game session optionally with a world seed
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSessionRequest"
      responses:
        "200":
          description: New session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionLog"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    SessionLog:
      type: object
      properties:
        sessionId:
          type: string
        turns:
          type: array
          items:
            $ref: "#/components/schemas/Turn"
        startedAt:
          type: string
        worldName:
          type: string
      required:
        - sessionId
        - turns
        - startedAt
        - worldName
    Turn:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        turnNumber:
          type: integer
        playerAction:
          type: string
        narrative:
          type: string
        worldUpdates:
          type: object
          additionalProperties: true
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - turnNumber
        - playerAction
        - narrative
        - createdAt
    PlayerTurn:
      type: object
      properties:
        action:
          type: string
          description: The player's action or dialogue
        sessionId:
          type: string
          description: The session ID to continue
      required:
        - action
        - sessionId
    TurnResult:
      type: object
      properties:
        turn:
          $ref: "#/components/schemas/Turn"
        worldUpdates:
          type: object
          additionalProperties: true
        updatedEntities:
          type: array
          items:
            type: string
      required:
        - turn
        - updatedEntities
    NewSessionRequest:
      type: object
      properties:
        worldName:
          type: string
        playerCharacterName:
          type: string
        playerCharacterClass:
          type: string
        worldSeed:
          type: string
          description: Optional seed description for the world
...