# lib/api-zod/src/generated/types/newSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RPG World Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface NewSessionRequest {
  worldName?: string;
  playerCharacterName?: string;
  playerCharacterClass?: string;
  /** Optional seed description for the world */
  worldSeed?: string;
}