# lib/api-zod/src/generated/types/playerTurn.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RPG World Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerTurn {
  /** The player's action or dialogue */
  action: string;
  /** The session ID to continue */
  sessionId: string;
}